Routify index.js
