+++ api/index.js
@@ -1,9 +1,7 @@
 const fs = require('fs');
 const jwt = require('jsonwebtoken');
 const Cacher = require('./lib/cacher');
-const Miss = Cacher.Miss;
 const { ValidationError } = require('express-json-validator-middleware');
-const Busboy = require('busboy');
 const Analytics = require('./lib/analytics');
 const path = require('path');
 const morgan = require('morgan');
@@ -20,15 +18,7 @@ const args = require('minimist')(process.argv, {
     string: ['postgres']
 });
 
-const Map = require('./lib/map');
 const Err = require('./lib/error');
-const Run = require('./lib/run');
-const Job = require('./lib/job');
-const JobError = require('./lib/joberror');
-const Data = require('./lib/data');
-const Upload = require('./lib/upload');
-const Schedule = require('./lib/schedule');
-const Collection = require('./lib/collections');
 
 const Param = util.Param;
 
@@ -65,9 +55,6 @@ function configure(args, cb) {
  */
 
 async function server(args, config, cb) {
-    // these must be run after lib/config
-    const ci = new (require('./lib/ci'))(config);
-
     let tb = false;
     if (!args['no-tilebase']) {
         console.log(`ok - loading: s3://${config.Bucket}/${config.StackName}/fabric.tilebase`);
@@ -113,6 +100,7 @@ async function server(args, config, cb) {
 
     config.cacher = new Cacher(args['no-cache']);
     config.pool = pool;
+    config.tb = tb;
 
     try {
         if (args.populate) {
@@ -250,46 +238,6 @@ async function server(args, config, cb) {
         if (!config.silent) console.error(`ok - loaded routes/${r}`);
         await require('./routes/' + r)(schema, config);
     }
-
-    /**
-     * @api {post} /api/upload Create Upload
-     * @apiVersion 1.0.0
-     * @apiName upload
-     * @apiGroup Upload
-     * @apiPermission upload
-     *
-     * @apiDescription
-     *     Statically cache source data
-     *
-     *     If a source is unable to be pulled from directly, authenticated users can cache
-     *     data resources to the OpenAddresses S3 cache to be pulled from
-     */
-    await schema.post('/upload', null,
-        async (req, res) => {
-            try {
-                await user.is_flag(req, 'upload');
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-
-            const busboy = new Busboy({ headers: req.headers });
-
-            const files = [];
-
-            busboy.on('file', (fieldname, file, filename) => {
-                files.push(Upload.put(req.auth.uid, filename, file));
-            });
-
-            busboy.on('finish', async () => {
-                try {
-                    res.json(await Promise.all(files));
-                } catch (err) {
-                    Err.respond(res, err);
-                }
-            });
-
-            return req.pipe(busboy);
-        });
 
     /**
      * @api {get} /api/user List Users
@@ -645,1178 +593,6 @@ async function server(args, config, cb) {
             }
         }
     );
-
-    /**
-     * @api {post} /api/schedule Scheduled Event
-     * @apiVersion 1.0.0
-     * @apiName Schedule
-     * @apiGroup Schedule
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *     Internal function to allow scheduled lambdas to kick off events
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.Schedule.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Standard.json} apiSuccess
-     */
-    await schema.post('/schedule', {
-        body: 'req.body.Schedule.json',
-        res: 'res.Standard.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            await Schedule.event(pool, req.body);
-
-            return res.json({
-                status: 200,
-                message: 'Schedule Event Started'
-            });
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/collections List Collections
-     * @apiVersion 1.0.0
-     * @apiName ListCollections
-     * @apiGroup Collections
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return a list of all collections and their glob rules
-     *
-     * @apiSchema {jsonawait schema=./schema/res.ListCollections.json} apiSuccess
-     */
-    await schema.get('/collections', {
-        res: 'res.ListCollections.json'
-    }, async (req, res) => {
-        try {
-            const collections = await config.cacher.get(Miss(req.query, 'collection'), async () => {
-                return await Collection.list(pool);
-            });
-
-            if (!req.auth || !req.auth.level || req.auth.level !== 'sponsor') {
-                for (const c of collections) {
-                    delete c.s3;
-                }
-            }
-
-            return res.json(collections);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/collections/:collection/data Get Collection Data
-     * @apiVersion 1.0.0
-     * @apiName DataCollection
-     * @apiGroup Collections
-     * @apiPermission user
-     *
-     * @apiDescription
-     *   Download a given collection file
-     *
-     *    Note: the user must be authenticated to perform a download. One of our largest costs is
-     *    S3 egress, authenticated downloads allow us to prevent abuse, keep the project running and the data free.
-     *
-     *    Faster Downloads? Have AWS? The Jobs, Data, & Collections API all return an `s3` property which links
-     *    to a requester pays object on S3. For those that are able, this is the best way to download data.
-     *
-     *    OpenAddresses is entirely funded by volunteers (many of them the developers themselves!)
-     *    Please consider donating if you are able https://opencollective.com/openaddresses
-     *
-     * @apiParam {Number} :collection Collection ID
-     */
-    await schema.get('/collections/:collection/data', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'collection');
-
-                await user.is_auth(req, true);
-
-                Collection.data(pool, req.params.collection, res);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {delete} /api/collections/:collection Delete Collection
-     * @apiVersion 1.0.0
-     * @apiName DeleteCollection
-     * @apiGroup Collections
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Delete a collection (This should not be done lightly)
-     *
-     * @apiParam {Number} :collection Collection ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.Standard.json} apiSuccess
-     */
-    await schema.delete('/collections/:collection', {
-        res: 'res.Standard.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'collection');
-
-            await user.is_admin(req);
-
-            await Collection.delete(pool, req.params.collection);
-
-            return res.json({
-                status: 200,
-                message: 'Collection Deleted'
-            });
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {post} /api/collections Create Collection
-     * @apiVersion 1.0.0
-     * @apiName CreateCollection
-     * @apiGroup Collections
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Create a new collection
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.CreateCollection.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Collection.json} apiSuccess
-     */
-    await schema.post('/collections', {
-        body: 'req.body.CreateCollection.json',
-        res: 'res.Collection.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            const collection = new Collection(req.body.name, req.body.sources);
-            await collection.generate(pool);
-
-            await config.cacher.del('collection');
-            return res.json(collection.json());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {patch} /api/collections/:collection Patch Collection
-     * @apiVersion 1.0.0
-     * @apiName PatchCollection
-     * @apiGroup Collections
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Update a collection
-     *
-     * @apiParam {Number} :collection Collection ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.PatchCollection.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Collection.json} apiSuccess
-     */
-    await schema.patch('/collections/:collection', {
-        body: 'req.body.PatchCollection.json',
-        res: 'res.Collection.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'collection');
-
-            await user.is_admin(req);
-
-            const collection = await Collection.from(pool, req.params.collection);
-
-            collection.patch(req.body);
-
-            await collection.commit(pool);
-            await config.cacher.del('collection');
-
-            return res.json(collection.json());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/map Coverage TileJSON
-     * @apiVersion 1.0.0
-     * @apiName TileJSON
-     * @apiGroup Map
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Data required for map initialization
-     */
-    await schema.get('/map', null,
-        (req, res) => {
-            return res.json(Map.map());
-        });
-
-    /**
-     * @api {get} /api/map/borders/:z/:x/:y.mvt Borders MVT
-     * @apiVersion 1.0.0
-     * @apiName BorderVectorTile
-     * @apiGroup Map
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Retrive borders Mapbox Vector Tiles
-     *
-     * @apiParam {Number} z Z coordinate
-     * @apiParam {Number} x X coordinate
-     * @apiParam {Number} y Y coordinate
-     */
-    await schema.get('/map/borders/:z/:x/:y.mvt', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'z');
-                await Param.int(req, 'x');
-                await Param.int(req, 'y');
-
-                if (req.params.z > 5) throw new Error(400, null, 'Up to z5 is supported');
-
-                const tile = await config.cacher.get(Miss(req.query, `tile-borders-${req.params.z}-${req.params.x}-${req.params.y}`), async () => {
-                    return await Map.border_tile(pool, req.params.z, req.params.x, req.params.y);
-                }, false);
-
-                res.type('application/vnd.mapbox-vector-tile');
-
-                return res.send(tile);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/map/fabric/:z/:x/:y.mvt Fabric MVT
-     * @apiVersion 1.0.0
-     * @apiName FabricVectorTile
-     * @apiGroup Map
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Retrive fabric Mapbox Vector Tiles
-     *
-     * @apiParam {Number} z Z coordinate
-     * @apiParam {Number} x X coordinate
-     * @apiParam {Number} y Y coordinate
-     */
-    await schema.get('/map/fabric/:z/:x/:y.mvt', null,
-        async (req, res) => {
-            await Param.int(req, 'z');
-            await Param.int(req, 'x');
-            await Param.int(req, 'y');
-
-            let tile;
-            try {
-                tile = await config.cacher.get(Miss(req.query, `tile-fabric-${req.params.z}-${req.params.x}-${req.params.y}`), async () => {
-                    return await Map.fabric_tile(tb, req.params.z, req.params.x, req.params.y);
-                }, false);
-
-                if (tile.length === 0) {
-                    throw new Err(404, null, 'No Tile Found');
-                }
-            } catch (err) {
-                return Err.respond(new Err(404, err, 'No Tile Found'), res);
-            }
-
-            res.writeHead(200, {
-                'Content-Type': 'application/vnd.mapbox-vector-tile',
-                'Content-Encoding': 'gzip'
-            });
-            res.end(tile);
-        });
-
-    /**
-     * @api {get} /api/map/:z/:x/:y.mvt Coverage MVT
-     * @apiVersion 1.0.0
-     * @apiName VectorTile
-     * @apiGroup Map
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Retrive coverage Mapbox Vector Tiles
-     *
-     * @apiParam {Number} z Z coordinate
-     * @apiParam {Number} x X coordinate
-     * @apiParam {Number} y Y coordinate
-     */
-    await schema.get('/map/:z/:x/:y.mvt', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'z');
-                await Param.int(req, 'x');
-                await Param.int(req, 'y');
-
-                if (req.params.z > 5) throw new Error(400, null, 'Up to z5 is supported');
-
-                const tile = await config.cacher.get(Miss(req.query, `tile-${req.params.z}-${req.params.x}-${req.params.y}`), async () => {
-                    return await Map.tile(pool, req.params.z, req.params.x, req.params.y);
-                }, false);
-
-                res.type('application/vnd.mapbox-vector-tile');
-
-                return res.send(tile);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/data List Data
-     * @apiVersion 1.0.0
-     * @apiName ListData
-     * @apiGroup Data
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Get the latest successful run of a given geographic area
-     *
-     * @apiSchema (Query) {jsonawait schema=./schema/req.query.ListData.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.ListData.json} apiSuccess
-     */
-    await schema.get('/data', {
-        query: 'req.query.ListData.json',
-        res: 'res.ListData.json'
-    }, async (req, res) => {
-        try {
-            const data = await config.cacher.get(Miss(req.query, 'data'), async () => {
-                return await Data.list(pool, req.query);
-            });
-
-            if (!req.auth || !req.auth.level || req.auth.level !== 'sponsor') {
-                for (const d of data) {
-                    delete d.s3;
-                }
-            }
-
-            return res.json(data);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {patch} /api/data/:data Update Data
-     * @apiVersion 1.0.0
-     * @apiName Update
-     * @apiGroup Data
-     * @apiPermission data
-     *
-     * @apiDescription
-     *   Update an existing data object
-     *
-     * @apiParam {Number} :data Data ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.PatchData.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Data.json} apiSuccess
-     *
-     */
-    await schema.patch('/data/:data', {
-        body: 'req.body.PatchData.json',
-        res: 'res.Data.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'data');
-
-            await user.is_admin(req);
-
-            req.body.id = req.params.data;
-
-            await config.cacher.del('data');
-
-            return res.json(await Data.commit(pool, req.body));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/data/:data Get Data
-     * @apiVersion 1.0.0
-     * @apiName SingleData
-     * @apiGroup Data
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return all information about a specific data segment
-     *
-     * @apiParam {Number} :data Data ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.Data.json} apiSuccess
-     */
-    await schema.get('/data/:data', {
-        res: 'res.Data.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'data');
-
-            const data = await Data.from(pool, req.params.data);
-
-            return res.json(data);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/data/:data/history Return Data History
-     * @apiVersion 1.0.0
-     * @apiName SingleHistoryData
-     * @apiGroup Data
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return the job history for a given data component
-     *
-     * @apiParam {Number} :data Data ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.DataHistory.json} apiSuccess
-     */
-    await schema.get('/data/:data/history', {
-        res: 'res.DataHistory.json'
-    }, async (req, res) => {
-        try {
-            const history = await Data.history(pool, req.params.data);
-
-            return res.json(history);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/run List Runs
-     * @apiVersion 1.0.0
-     * @apiName ListRuns
-     * @apiGroup Run
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Runs are container objects that contain jobs that were started at the same time or by the same process
-     *
-     * @apiParam {Number} :data Data ID
-     *
-     * @apiSchema (Query) {jsonawait schema=./schema/req.query.ListRuns.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.ListRuns.json} apiSuccess
-     */
-    await schema.get('/run', {
-        query: 'req.query.ListRuns.json',
-        res: 'res.ListRuns.json'
-    }, async (req, res) => {
-        try {
-            if (req.query.status) req.query.status = req.query.status.split(',');
-            const runs = await Run.list(pool, req.query);
-
-            return res.json(runs);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {post} /api/run Create Run
-     * @apiVersion 1.0.0
-     * @apiName CreateRun
-     * @apiGroup Run
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Create a new run to hold a batch of jobs
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.CreateRun.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Run.json} apiSuccess
-     */
-    await schema.post('/run', {
-        body: 'req.body.CreateRun.json',
-        res: 'res.Run.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            const run = await Run.generate(pool, req.body);
-
-            return res.json(run.json());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/run/:run Get Run
-     * @apiVersion 1.0.0
-     * @apiName Single
-     * @apiGroup Run
-     * @apiPermission public
-     *
-     * @apiParam {Number} :run Run ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.Run.json} apiSuccess
-     */
-    await schema.get('/run/:run', {
-        res: 'res.Run.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'run');
-
-            res.json(await Run.from(pool, req.params.run));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/run/:run/count Run Stats
-     * @apiVersion 1.0.0
-     * @apiName RunStats
-     * @apiGroup Run
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return statistics about jobs within a given run
-     *
-     * @apiParam {Number} :run Run ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.RunStats.json} apiSuccess
-     */
-    await schema.get('/run/:run/count', {
-        res: 'res.RunStats.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'run');
-
-            res.json(await Run.stats(pool, req.params.run));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {patch} /api/run/:run Update Run
-     * @apiVersion 1.0.0
-     * @apiName Update
-     * @apiGroup Run
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Update an existing run
-     *
-     * @apiParam {Number} :run Run ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.PatchRun.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Run.json} apiSuccess
-     *
-     */
-    await schema.patch('/run/:run', {
-        body: 'req.body.PatchRun.json',
-        res: 'res.Run.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'run');
-
-            await user.is_admin(req);
-
-            const run = await Run.from(pool, req.params.run);
-
-            // The CI is making a CI run "live" and updating the /data list
-            if ((!run.live && req.body.live) || (run.live && !req.body.live)) await config.cacher.del('data');
-
-            run.patch(req.body);
-
-            await run.commit(pool);
-
-            return res.json(run.json());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {post} /api/run/:run/jobs Populate Run Jobs
-     * @apiVersion 1.0.0
-     * @apiName SingleJobsCreate
-     * @apiGroup Run
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *     Given an array sources, explode it into multiple jobs and submit to batch
-     *     or pass in a predefined list of sources/layer/names
-     *
-     *     Note: once jobs are attached to a run, the run is "closed" and subsequent
-     *     jobs cannot be attached to it
-     *
-     * @apiParam {Number} :run Run ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.SingleJobsCreate.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.SingleJobsCreate.json} apiSuccess
-     */
-    await schema.post('/run/:run/jobs', {
-        body: 'req.body.SingleJobsCreate.json',
-        res: 'res.SingleJobsCreate.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'run');
-
-            await user.is_admin(req);
-
-            return res.json(await Run.populate(pool, req.params.run, req.body.jobs));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/run/:run/jobs List Run Jobs
-     * @apiVersion 1.0.0
-     * @apiName SingleJobs
-     * @apiGroup Run
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return all jobs for a given run
-     *
-     * @apiParam {Number} :run Run ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.SingleJobs.json} apiSuccess
-     */
-    await schema.get('/run/:run/jobs', {
-        res: 'res.SingleJobs.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'run');
-
-            const jobs = await Run.jobs(pool, req.params.run);
-
-            res.json({
-                run: req.params.run,
-                jobs: jobs
-            });
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job List Jobs
-     * @apiVersion 1.0.0
-     * @apiName ListJobs
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return information about a given subset of jobs
-     *
-     * @apiSchema (query) {jsonawait schema=./schema/req.query.ListJobs.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.ListJobs.json} apiSuccess
-     */
-    await schema.get('/job', {
-        query: 'req.query.ListJobs.json',
-        res: 'res.ListJobs.json'
-    }, async (req, res) => {
-        try {
-            if (req.query.status) req.query.status = req.query.status.split(',');
-
-            const jobs = await Job.list(pool, req.query);
-
-            if (!req.auth || !req.auth.level || req.auth.level !== 'sponsor') {
-                for (const j of jobs) {
-                    delete j.s3;
-                }
-            }
-
-            return res.json(jobs);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/error Get Job Errors
-     * @apiVersion 1.0.0
-     * @apiName ErrorList
-     * @apiGroup JobErrors
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     All jobs that fail as part of a live run are entered into the JobError API
-     *     This API powers a page that allows for human review of failing jobs
-     *     Note: Job Errors are cleared with every subsequent full cache
-     *
-     * @apiSchema {jsonawait schema=./schema/res.ErrorList.json} apiSuccess
-     */
-    await schema.get('/job/error', {
-        res: 'res.ErrorList.json'
-    }, async (req, res) => {
-        try {
-            return res.json(await JobError.list(pool, req.query));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/error/count Job Error Count
-
-     * @apiVersion 1.0.0
-     * @apiName ErrorCount
-     * @apiGroup JobError
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return a simple count of the current number of job errors
-     *
-     * @apiSchema {jsonawait schema=./schema/res.ErrorCount.json} apiSuccess
-     */
-    await schema.get('/job/error/count', {
-        res: 'res.ErrorCount.json'
-    }, async (req, res) => {
-        try {
-            return res.json(await JobError.count(pool));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-
-    /**
-     * @api {get} /api/job/error/:job Get Job Error
-     * @apiVersion 1.0.0
-     * @apiName ErrorList
-     * @apiGroup ErrorSingle
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return a single job error if one exists or 404 if not
-     *
-     * @apiSchema {jsonawait schema=./schema/res.ErrorSingle.json} apiSuccess
-     */
-    await schema.get('/job/error/:job', {
-        res: 'res.ErrorSingle.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            return res.json(await JobError.get(pool, req.params.job));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {post} /api/job/error Create Job Error
-     * @apiVersion 1.0.0
-     * @apiName ErrorCreate
-     * @apiGroup JobError
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *     Create a new Job Error in response to a live job that Failed or Warned
-     *
-     * @apiParam {Number} job Job ID of the given error
-     * @apiParam {String} message Text representation of the error
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.ErrorCreate.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.ErrorCreate.json} apiSuccess
-     */
-    await schema.post('/job/error', {
-        body: 'req.body.ErrorCreate.json',
-        res: 'res.ErrorCreate.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            const joberror = new JobError(req.body.job, req.body.message);
-            return res.json(await joberror.generate(pool));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {post} /api/job/error/:job Resolve Job Error
-     * @apiVersion 1.0.0
-     * @apiName ErrorModerate
-     * @apiGroup JobError
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *     Mark a job error as resolved
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/res.ErrorModerate.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.ErrorModerate.json} apiSuccess
-     */
-    await schema.post('/job/error/:job', {
-        body: 'req.body.ErrorModerate.json',
-        res: 'res.ErrorModerate.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            await user.is_flag(req, 'moderator');
-
-            res.json(await JobError.moderate(pool, ci, req.params.job, req.body));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/:job Get Job
-     * @apiVersion 1.0.0
-     * @apiName Single
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return all information about a given job
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.Job.json} apiSuccess
-     */
-    await schema.get('/job/:job', {
-        res: 'res.Job.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            const job = (await Job.from(pool, req.params.job)).json();
-
-            if (!req.auth || !req.auth.level || req.auth.level !== 'sponsor') {
-                delete job.s3;
-            }
-
-            return res.json(job);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/:job/raw Raw Source
-     * @apiVersion 1.0.0
-     * @apiName RawSingle
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *     Return the raw source from github - this API is not stable nor
-     *     will it always return a consistent result
-     *
-     * @apiParam {Number} :job Job ID
-     */
-    await schema.get('/job/:job/raw', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'job');
-
-                const job = await Job.from(pool, req.params.job);
-
-                return res.json(await job.get_raw());
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {post} /api/job/:job Rerun Job
-     * @apiVersion 1.0.0
-     * @apiName JobRerun
-     * @apiGroup Job
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *     Submit a job for reprocessing - often useful for network errors
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.SingleJobsCreate.json} apiSuccess
-     */
-    await schema.post('/job/:job/rerun', {
-        res: 'res.SingleJobsCreate.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            await user.is_admin(req);
-
-            const job = await Job.from(pool, req.params.job);
-            const run = await Run.from(pool, job.run);
-
-            const new_run = await Run.generate(pool, {
-                live: !!run.live
-            });
-
-            return res.json(await Run.populate(pool, new_run.id, [{
-                source: job.source,
-                layer: job.layer,
-                name: job.name
-            }]));
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/:job/delta Job Stats Comparison
-     * @apiVersion 1.0.0
-     * @apiName SingleDelta
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Compare the stats of the given job against the current live data job
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.SingleDelta.json} apiSuccess
-     */
-    await schema.get('/job/:job/delta', {
-        res: 'res.SingleDelta.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            const delta = await Job.delta(pool, req.params.job);
-
-            return res.json(delta);
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/job/:job/output/source.png Get Job Preview
-     * @apiVersion 1.0.0
-     * @apiName SingleOutputPreview
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return the preview image for a given job
-     *
-     * @apiParam {Number} :job Job ID
-     */
-    await schema.get('/job/:job/output/source.png', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'job');
-                Job.preview(req.params.job, res);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/job/:job/output/source.geojson.gz Get Job Data
-     * @apiVersion 1.0.0
-     * @apiName SingleOutputData
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *    Note: the user must be authenticated to perform a download. One of our largest costs is
-     *    S3 egress, authenticated downloads allow us to prevent abuse and keep the project running and the data freetw
-     *
-     *    Faster Downloads? Have AWS? The Jobs, Data, & Collections API all return an `s3` property which links
-     *    to a requester pays object on S3. For those that are able, this is the best way to download data.
-     *
-     *    OpenAddresses is entirely funded by volunteers (many of then the developers themselves!)
-     *    Please consider donating if you are able https://opencollective.com/openaddresses
-     *
-     * @apiParam {Number} :job Job ID
-     */
-    await schema.get('/job/:job/output/source.geojson.gz', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'job');
-
-                await user.is_auth(req, true);
-
-                await Job.data(pool, req.params.job, res);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/job/:job/output/sample Small Sample
-     * @apiVersion 1.0.0
-     * @apiName SampleData
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return an Array containing a sample of the properties
-     *
-     * @apiParam {Number} :job Job ID
-     */
-    await schema.get('/job/:job/output/sample', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'job');
-
-                return res.json(await Job.sample(pool, req.params.job));
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/job/:job/output/cache.zip Get Job Cache
-     * @apiVersion 1.0.0
-     * @apiName SingleOutputCache
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     *  @apiDescription
-     *    Note: the user must be authenticated to perform a download. One of our largest costs is
-     *    S3 egress, authenticated downloads allow us to prevent abuse and keep the project running and the data freetw
-     *
-     *    Faster Downloads? Have AWS? The Jobs, Data, & Collections API all return an `s3` property which links
-     *    to a requester pays object on S3. For those that are able, this is the best way to download data.
-     *
-     *    OpenAddresses is entirely funded by volunteers (many of then the developers themselves!)
-     *    Please consider donating if you are able https://opencollective.com/openaddresses
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     */
-    await schema.get('/job/:job/output/cache.zip', null,
-        async (req, res) => {
-            try {
-                await Param.int(req, 'job');
-
-                await user.is_auth(req, true);
-
-                Job.cache(req.params.job, res);
-            } catch (err) {
-                return Err.respond(err, res);
-            }
-        });
-
-    /**
-     * @api {get} /api/job/:job/log Get Job Log
-     * @apiVersion 1.0.0
-     * @apiName SingleLog
-     * @apiGroup Job
-     * @apiPermission public
-     *
-     * @apiDescription
-     *   Return the batch-machine processing log for a given job
-     *   Note: These are stored in AWS CloudWatch and *do* expire
-     *   The presence of a loglink on a job, does not guarentree log retention
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema {jsonawait schema=./schema/res.SingleLog.json} apiSuccess
-     */
-    await schema.get('/job/:job/log', {
-        res: 'res.SingleLog.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            const job = await Job.from(pool, req.params.job);
-
-            return res.json(await job.log());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {patch} /api/job/:job Update Job
-     * @apiVersion 1.0.0
-     * @apiName JobPatch
-     * @apiGroup Job
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Update a job
-     *
-     * @apiParam {Number} :job Job ID
-     *
-     * @apiSchema (Body) {jsonawait schema=./schema/req.body.PatchJob.json} apiParam
-     * @apiSchema {jsonawait schema=./schema/res.Job.json} apiSuccess
-     */
-    await schema.patch('/job/:job', {
-        body: 'req.body.PatchJob.json',
-        res: 'res.Job.json'
-    }, async (req, res) => {
-        try {
-            await Param.int(req, 'job');
-
-            await user.is_admin(req);
-
-            const job = await Job.from(pool, req.params.job);
-            job.patch(req.body);
-            await job.commit(pool, Run, Data, ci);
-
-            await Run.ping(pool, ci, job);
-
-            return res.json(job.json());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/dash/traffic Session Counts
-     * @apiVersion 1.0.0
-     * @apiName TrafficAnalytics
-     * @apiGroup Analytics
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Report anonymized traffic data about the number of user sessions created in a given day.
-     *
-     * @apiSchema {jsonawait schema=./schema/res.TrafficAnalytics.json} apiSuccess
-     */
-    await schema.get('/dash/traffic', {
-        res: 'res.TrafficAnalytics.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            res.json(await analytics.traffic());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    /**
-     * @api {get} /api/dash/collections Collection Counts
-     * @apiVersion 1.0.0
-     * @apiName CollectionsAnalytics
-     * @apiGroup Analytics
-     * @apiPermission admin
-     *
-     * @apiDescription
-     *   Report anonymized traffic data about the number of collection downloads.
-     *
-     * @apiSchema {jsonawait schema=./schema/res.CollectionsAnalytics.json} apiSuccess
-     */
-    await schema.get('/dash/collections', {
-        res: 'res.CollectionsAnalytics.json'
-    }, async (req, res) => {
-        try {
-            await user.is_admin(req);
-
-            res.json(await analytics.collections());
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
 
     schema.router.use((err, req, res, next) => {
         if (err instanceof ValidationError) {
